Revert back to 32 bit inode values

+++ src/main/java/org/redline_rpm/payload/Contents.java
@@ -738,8 +738,8 @@ public class Contents {
 	 * Gets the inodes header values.
 	 * @return the iNodes header values
 	 */
-	public long[] getInodes() {
-		long[] array = new long[ headers.size()];
+	public int[] getInodes() {
+		int[] array = new int[ headers.size()];
 		int x = 0;
 		for ( CpioHeader header : headers) array[ x++] = header.getInode();
 		return array;
